+++ crates/krilla/src/object/mask.rs
@@ -155,33 +155,8 @@ mod tests {
         sc.register_cacheable(mask);
     }
 
-    fn mask_visreg_impl(mask_type: MaskType, surface: &mut Surface, color: rgb::Color) {
-        let path = rect_to_path(20.0, 20.0, 180.0, 180.0);
-        let mask = basic_mask(surface, mask_type);
-        surface.push_mask(mask);
-        surface.fill_path(
-            &path,
-            Fill {
-                paint: color.into(),
-                opacity: NormalizedF32::ONE,
-                rule: Default::default(),
-            },
-        );
-        surface.pop();
-    }
-
     #[snapshot]
     pub fn mask_luminosity(sc: &mut SerializeContext) {
         mask_snapshot_impl(MaskType::Luminosity, sc);
-    }
-
-    #[visreg(all)]
-    pub fn mask_luminosity(surface: &mut Surface) {
-        mask_visreg_impl(MaskType::Luminosity, surface, rgb::Color::new(0, 255, 0));
-    }
-
-    #[visreg(all)]
-    pub fn mask_alpha(surface: &mut Surface) {
-        mask_visreg_impl(MaskType::Luminosity, surface, rgb::Color::new(0, 0, 128));
     }
 }
